+++ android/build.gradle
@@ -13,7 +13,7 @@ buildscript {
         mavenCentral()
     }
     dependencies {
-        classpath("com.android.tools.build:gradle:4.2.2")
+        classpath('com.android.tools.build:gradle:7.1.2')
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
