+++ src/redux/configureStore.js
@@ -1,10 +1,17 @@
-import {combineReducers, createStore} from 'redux';
+import {combineReducers, createStore, applyMiddleware} from 'redux';
 import counterReducer from './ducks/counter';
+import userReducer from './ducks/user';
+import createSagaMiddleware from 'redux-saga';
+import {watcherSaga} from './sagas/rootSaga';
 
 const reducer = combineReducers({
   counter: counterReducer,
+  user: userReducer,
 });
 
-const store = createStore(reducer);
+const sagaMiddleware = createSagaMiddleware();
+const middleware = [sagaMiddleware];
+const store = createStore(reducer, {}, applyMiddleware(...middleware));
 
+sagaMiddleware.run(watcherSaga);
 export default store;
