+++ src/redux/sagas/requests/user.js
@@ -1,0 +1,5 @@
+import {api} from 'apis';
+
+export const requestGetUser = async () => {
+  return await api.get('/search/repositories?q=user:reactjs');
+};
