+++ src/redux/sagas/rootSaga.js
@@ -1,0 +1,7 @@
+import {takeLatest} from 'redux-saga/effects';
+import {handleGetUser} from './handler/user';
+import {GET_USER} from '../ducks/user';
+
+export function* watcherSaga() {
+  yield takeLatest(GET_USER, handleGetUser);
+}
