+++ tailwind.css
@@ -6,18 +6,6 @@
   bottom: 3rem
 }
 
-.left-0 {
-  left: 0px
-}
-
-.right-0 {
-  right: 0px
-}
-
-.right-12 {
-  right: 3rem
-}
-
 .right-3 {
   right: 0.75rem
 }
@@ -41,14 +29,6 @@
 
 .mb-5 {
   margin-bottom: 1.25rem
-}
-
-.mt-14 {
-  margin-top: 3.5rem
-}
-
-.mb-3 {
-  margin-bottom: 0.75rem
 }
 
 .flex {
@@ -87,11 +67,6 @@
   border-width: 1px
 }
 
-.border-gray-800 {
-  --tw-border-opacity: 1;
-  border-color: rgb(31 41 55 / var(--tw-border-opacity))
-}
-
 .bg-gray-300 {
   --tw-bg-opacity: 1;
   background-color: rgb(209 213 219 / var(--tw-bg-opacity))
@@ -102,32 +77,9 @@
   background-color: rgb(255 255 255 / var(--tw-bg-opacity))
 }
 
-.bg-black {
-  --tw-bg-opacity: 1;
-  background-color: rgb(0 0 0 / var(--tw-bg-opacity))
-}
-
 .bg-blue-200 {
   --tw-bg-opacity: 1;
   background-color: rgb(191 219 254 / var(--tw-bg-opacity))
-}
-
-.bg-gray-800 {
-  --tw-bg-opacity: 1;
-  background-color: rgb(31 41 55 / var(--tw-bg-opacity))
-}
-
-.bg-gray-400 {
-  --tw-bg-opacity: 1;
-  background-color: rgb(156 163 175 / var(--tw-bg-opacity))
-}
-
-.p-3 {
-  padding: 0.75rem
-}
-
-.p-10 {
-  padding: 2.5rem
 }
 
 .p-7 {
@@ -180,11 +132,6 @@
 .text-gray-600 {
   --tw-text-opacity: 1;
   color: rgb(75 85 99 / var(--tw-text-opacity))
-}
-
-.text-white {
-  --tw-text-opacity: 1;
-  color: rgb(255 255 255 / var(--tw-text-opacity))
 }
 
 @media (prefers-color-scheme: dark) {
