+++ tailwind.json
@@ -9,21 +9,6 @@
 			"bottom": 48
 		}
 	},
-	"left-0": {
-		"style": {
-			"left": 0
-		}
-	},
-	"right-0": {
-		"style": {
-			"right": 0
-		}
-	},
-	"right-12": {
-		"style": {
-			"right": 48
-		}
-	},
 	"right-3": {
 		"style": {
 			"right": 12
@@ -56,16 +41,6 @@
 	"mb-5": {
 		"style": {
 			"marginBottom": 20
-		}
-	},
-	"mt-14": {
-		"style": {
-			"marginTop": 56
-		}
-	},
-	"mb-3": {
-		"style": {
-			"marginBottom": 12
 		}
 	},
 	"flex": {
@@ -124,15 +99,6 @@
 			"borderLeftWidth": 1
 		}
 	},
-	"border-gray-800": {
-		"style": {
-			"--tw-border-opacity": 1,
-			"borderTopColor": "rgb(31 41 55 / var(--tw-border-opacity))",
-			"borderRightColor": "rgb(31 41 55 / var(--tw-border-opacity))",
-			"borderBottomColor": "rgb(31 41 55 / var(--tw-border-opacity))",
-			"borderLeftColor": "rgb(31 41 55 / var(--tw-border-opacity))"
-		}
-	},
 	"bg-gray-300": {
 		"style": {
 			"--tw-bg-opacity": 1,
@@ -145,44 +111,10 @@
 			"backgroundColor": "rgb(255 255 255 / var(--tw-bg-opacity))"
 		}
 	},
-	"bg-black": {
-		"style": {
-			"--tw-bg-opacity": 1,
-			"backgroundColor": "rgb(0 0 0 / var(--tw-bg-opacity))"
-		}
-	},
 	"bg-blue-200": {
 		"style": {
 			"--tw-bg-opacity": 1,
 			"backgroundColor": "rgb(191 219 254 / var(--tw-bg-opacity))"
-		}
-	},
-	"bg-gray-800": {
-		"style": {
-			"--tw-bg-opacity": 1,
-			"backgroundColor": "rgb(31 41 55 / var(--tw-bg-opacity))"
-		}
-	},
-	"bg-gray-400": {
-		"style": {
-			"--tw-bg-opacity": 1,
-			"backgroundColor": "rgb(156 163 175 / var(--tw-bg-opacity))"
-		}
-	},
-	"p-3": {
-		"style": {
-			"paddingTop": 12,
-			"paddingRight": 12,
-			"paddingBottom": 12,
-			"paddingLeft": 12
-		}
-	},
-	"p-10": {
-		"style": {
-			"paddingTop": 40,
-			"paddingRight": 40,
-			"paddingBottom": 40,
-			"paddingLeft": 40
 		}
 	},
 	"p-7": {
@@ -249,12 +181,6 @@
 		"style": {
 			"--tw-text-opacity": 1,
 			"color": "rgb(75 85 99 / var(--tw-text-opacity))"
-		}
-	},
-	"text-white": {
-		"style": {
-			"--tw-text-opacity": 1,
-			"color": "rgb(255 255 255 / var(--tw-text-opacity))"
 		}
 	},
 	"dark:bg-black": {
